Checking in 2 new utilities into the odie module.
Adding the codebale module into tcllib as there are a pile of routines we will probably want from there, and while
we are de-odie-er-izing code, we might as well get the whole thing.

(Checked into the feature branch first so I can have a version that arrives "as imported from odielib").