Added facilities for web pages to return a file instead of a dynamic buffer.
Eliminted the special "output" method for SCGI handlers. In truth, the only difference
was in the format of the first line, and I put together an "EncodeStatus" method 
which sorts that out.

We now change the translation mode of the channel during output to binary. I was getting strange
line break issues on the Mac without it.

Fixed OO::dialect such that Alias uses the same canonical mapping function as superclass.