Update the project.shed
The SCGI components of the httpd module now work well enough to operate a fossil repository

Revised the dispatch interaction between the http reply and the server. Dispatching now
takes place before MIME headers are read, based on the URI given in the GET/POST/etc

Added a reset method to dictobjs, as well as a behavior to inject the initial value
when replace is called.