Httpd improvements
All httpd replies are now given a guuid, and an object and coroutine name in the 
global namespace. Simplified error handling and detangled a web of try { ... }
that was confounding error reporting.

Added an "html" ensemble to server an reply object which take care of generating
pro-forma blocks of Html code, as well as provide shims for sitewide templating.

Reply dispatch methods no longer attempt their own error handling.

Going all-in with having coroutines monitor the entire lifecycle of a request