namespacex - Forward porting commit [ccd9433cb2abd51c] Part 1: Tests.
Extended the testsuite to demonstrate that the current command
implementations mishandle relative namespace arguments if they are not
refering to a child of the global namespace ::

Note, the changes made here diverge totally from the changes made in
the original commit. Explicitly added new tests to cover the
mishandled cases, leaving the existing coverage intact. The original
commit OTOH shifted the entire testsuite over to cover relative
namespace names, retracting the coverage for fully-qualified namespace
names.