namespacex - Fill in
Extended testsuite to cover new commands `normalize` and `strip`.
Extended testsuite to cover wrong#args for `import`.

Fixed typo in documentation of `strip`.

Split `strip` into public and internal forms, with the public form
performing argument normalization and checking not required by the
internal form (*), and documented for the public.

(*) The internal form has only a minimal check asserting that the
    prefix namespace is given as an FQN.