Added a new channel registration system for httpd replies, instead of forcing developers to manually clean up channels via the old TransferComplete mechanism.
Rather than hard-calling destructors for httpd reply objects (and their proxies)
added a soft-kill mechanism developed for cuneiform to the httpd module. A new
command "clay::destroy" marks and object for death, to be cleaned up at the 
next call for clay::cleanup. The cleanup is invoked at points in the idle
loop where the object is expected to be long done with its toys.

When an object wants to deliver its contents now, it can invoke the "close"
method. Close will now close ALL channels registered with ChannelRegister,
not just the channel in the "chan" variable.

Fixes the crashes in scgi test 0006, but now the content is still coming out
"wrong"