mime - I, Bump to 1.6.3.
Renamed a few placeholder arguments in the md5, etc. emulations. Triggered for `--` to avoid clash with a variant definition of `proc` (at FlightAware) which gives `--` special syntactic meaning in support of named parameter handling. The new name also makes it clear that these arguments are not used.

No functional changes.

Doc refresh.